+++ src/components/appHeader/AppWindowButtons.vue
@@ -2,7 +2,7 @@
 
    <q-btn-group
       flat
-      class="appWindowButtons"
+      class="appWindowButtons bg-dark"
    >
 
     <projectCloseCheckDialog
@@ -106,12 +106,14 @@ export default class AppWindowButtons extends BaseClass {
 
 <style lang="scss" scoped>
 .appWindowButtons {
+  border-radius: 0;
   position: fixed;
   right: 0;
   top: 0;
   height: 40px;
   z-index: 99999999;
   color: #fff;
+  -webkit-app-region: no-drag;
 }
 </style>