+++ src/components/fields/Field_SingleRelationship.vue
@@ -39,6 +39,10 @@
 
   <div class="flex" v-if="editMode">
     <q-select
+      class="singleRelashionshipSelect"
+      menu-anchor="bottom middle"
+      menu-self="top middle"
+      dark
       style="flex-grow: 1;"
       dense
       :options="filteredInput"
@@ -55,7 +59,10 @@
             v-on="itemEvents"
           >
             <q-item-section>
-              <q-item-label v-html="opt.label" ></q-item-label>
+              <q-item-label
+              :style="`color: ${opt.color}`"
+               v-html="opt.label" ></q-item-label>
+              <q-item-label caption class="text-cultured">{{opt.hierarchicalPath}}</q-item-label>
             </q-item-section>
             <q-tooltip v-if='opt.disable'>
               This option is unavailable for selection as it is already paired to another.
@@ -187,7 +194,9 @@ export default class Field_SingleRelationship extends BaseClass {
     if (this.inputDataBluePrint?.relationshipSettings && this.currentId.length > 0) {
       const CurrentObjectDB = new PouchDB(this.inputDataBluePrint.relationshipSettings.connectedObjectType)
 
-      const allObjects = (await CurrentObjectDB.allDocs({ include_docs: true })).rows.map((doc) => {
+      const allDbObjects = (await CurrentObjectDB.allDocs({ include_docs: true })).rows
+
+      const allObjects = allDbObjects.map((doc) => {
         const objectDoc = doc.doc as unknown as I_ShortenedDocument
 
         const pairedField = (this.inputDataBluePrint?.relationshipSettings?.connectedField) || ""
@@ -210,8 +219,11 @@ export default class Field_SingleRelationship extends BaseClass {
           disable: isDisabled,
           url: `/project/display-content/${objectDoc.type}/${objectDoc._id}`,
           label: objectDoc.extraFields.find(e => e.id === "name")?.value,
+          color: objectDoc.extraFields.find(e => e.id === "documentColor")?.value,
           isCategory: objectDoc.extraFields.find(e => e.id === "categorySwitch")?.value,
-          pairedField: pairedField
+          pairedField: pairedField,
+          // @ts-ignore
+          hierarchicalPath: this.getDocumentHieararchicalPath(objectDoc, allDbObjects)
         }
       }) as unknown as I_FieldRelationship[]