+++ src/store/module-openedDocuments/state.ts
@@ -3,6 +3,7 @@ import { I_OpenedDocument } from "./../../interfaces/I_OpenedDocument"
 export interface OpenDocumentsStateInterface {
   documents: {
     timestamp: string,
+    treeAction: boolean,
     docs: I_OpenedDocument[]
   }
 }
@@ -11,6 +12,7 @@ function state (): OpenDocumentsStateInterface {
   return {
     documents: {
       timestamp: "",
+      treeAction: false,
       docs: []
     }
   }